Applies normalization methods to given data set

Applies single or multiple normalization methods to the given data. Available methods are:

Total Count (TC)

Upper Quartile (UQ)

Median (median)

Trimmed Median of Means (TMM)

DESeq

PoissonSeq

Quantile Normalization (QN)

Remove Unwanted Variation (RUVg, RUVr, and RUVs)

Usage

applyNormalization(
  data,
  groups,
  method = c("TC", "UQ", "median", "TMM", "DESeq", "PoissonSeq", "QN", "RUV")
)

Arguments

data

Un-normalized count data set of shape p x n, where p is the number of markers n is the number of samples.

groups

Vector of length n that maps the samples to sample groups.

method

Vector of normalization methods that are applied to the data. Available methods are: c("TC", "UQ", "median", "TMM", "DESeq", "PoissonSeq", "QN", "RUV"). Select one or multiple methods. By default all normalization methods will be applied.

Value

List of normalized read counts for selected normalization methods.